rename : 디렉토리 명 shDto로 변경
