+++ server/src/main/java/com/example/hackathon/common/shDto/SHApiRECResponse.java
@@ -1,4 +1,4 @@
-package com.example.hackathon.common.shResponse;
+package com.example.hackathon.common.shDto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Builder;
