+++ server/src/main/java/com/example/hackathon/common/shDto/SHHeader.java
@@ -1,4 +1,4 @@
-package com.example.hackathon.common.shResponse;
+package com.example.hackathon.common.shDto;
 
 import com.example.hackathon.common.TransactionUniqueNoGenerator;
 import lombok.Builder;
